requirements.txt
numpy
pandas
matplotlib
scikit-learn
opencv-python
pillow

tensorflow
keras

fastapi
uvicorn[standard]
python-multipart

pydantic
requests
python-dotenv